requirements.txt
ossapi
python-dotenv
requests
